refactor: 💡 rename tooltippopup to popover Authors: Gustav Pettersson, Marcus Hogler
