--- src/common/TooltipPopup/index.ts
@@ -1,1 +1,0 @@
-export { default as TooltipPopup } from './TooltipPopup';
